fix(P3): Resolve config cache races, motion lock mixing, API limits, and hardcoded credentials
**Issue 1: Config cache write synchronization**
- configSetInt/Float/String updated config_table without mutex protection
- Solution: Added config_cache_mutex protection to all three setter functions
- Files: src/config_unified.cpp (lines 488-636)

**Issue 2: Motion state lock domain mixing**
- m_state reads used spinlock while some writes used motion mutex
- Caused potential torn/stale reads under contention
- Solution: Unified all m_state accesses to use motionSpinlock consistently
- Files: src/motion_control.cpp
  * motionHome() (lines 680-694)
  * motionMoveTo() (lines 776-780)
  * motionStop() (lines 939-942)
  * motionPause() (lines 969-972)
  * motionResume() (lines 1005-1008)
  * motionDwell() (lines 1042-1055)
  * motionWaitPin() (lines 1082-1098)

**Issue 3: Config API unsigned soft limits**
- API enforced min=0 via uint16_t, blocking negative coordinates
- validateSoftLimit() already supported -10000 to +10000 but was bypassed
- Solution:
  * Changed motion_config_t arrays from uint16_t to int32_t
  * Changed value.as<uint16_t>() to value.as<int32_t>() (6 locations)
  * Updated schema min from 0 to -10000 for all soft limit fields
- Files:
  * include/api_config.h (lines 35-37)
  * src/api_config.cpp (lines 228-240, 325-350)

**Issue 4: Web UI hardcoded Basic Auth credentials**
- gcode.js and safety.js had 'admin:password' in Authorization headers
- Solution: Removed hardcoded credentials, browser now handles auth via 401 prompt
- Files:
  * data/pages/gcode/gcode.js (line 183-190)
  * data/shared/safety.js (lines 88-91, 193-197, 219-223)
- Added credentials: 'same-origin' to fetch calls for proper auth handling

All changes marked with PHASE 5.10 comments for traceability.

+++ include/api_config.h
@@ -30,10 +30,11 @@ typedef enum {
 
 /**
  * @brief Motion configuration structure
+ * @note PHASE 5.10: Changed from uint16_t to int32_t to support negative coordinates
  */
 typedef struct {
-    uint16_t soft_limit_low_mm[3];      // Lower limit for X, Y, Z (mm)
-    uint16_t soft_limit_high_mm[3];     // Upper limit for X, Y, Z (mm)
+    int32_t soft_limit_low_mm[3];       // Lower limit for X, Y, Z (mm) - supports negative
+    int32_t soft_limit_high_mm[3];      // Upper limit for X, Y, Z (mm)
 } motion_config_t;
 
 /**
